netlify/functions/requirements.txt
# Netlify Python function deps (copied from api/requirements.txt)
redis==5.3.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
pymongo==4.9.1
ollama==0.3.3
groq==0.4.1
sentence-transformers==3.1.1
beautifulsoup4==4.12.3
bleach==6.1.0
html2text==2020.1.16
bs4==0.0.1
python-multipart==0.0.12
email-validator==2.2.0
